+++ contracts/test/TokenConservationWrapper.sol
@@ -4,17 +4,17 @@ import "../libraries/TokenConservation.sol";
 
 
 contract TokenConservationWrapper {
-    using TokenConservation for int[];
+    using TokenConservation for int256[];
     using TokenConservation for uint16[];
 
     function updateTokenConservationTest(
-        int[] memory testInstance,
+        int256[] memory testInstance,
         uint16 buyToken,
         uint16 sellToken,
         uint16[] memory tokenIdsForPrice,
         uint128 buyAmount,
         uint128 sellAmount
-    ) public pure returns(int[] memory) {
+    ) public pure returns (int256[] memory) {
         testInstance.updateTokenConservation(
             buyToken,
             sellToken,
@@ -25,15 +25,18 @@ contract TokenConservationWrapper {
         return testInstance;
     }
 
-    function checkTokenConservationTest(
-        int[] memory testInstance
-    ) public pure {
+    function checkTokenConservationTest(int256[] memory testInstance)
+        public
+        pure
+    {
         testInstance.checkTokenConservation();
     }
 
-    function checkPriceOrdering(
-        uint16[] memory tokenIdsForPrice
-    ) public pure returns(bool) {
+    function checkPriceOrdering(uint16[] memory tokenIdsForPrice)
+        public
+        pure
+        returns (bool)
+    {
         return tokenIdsForPrice.checkPriceOrdering();
     }
 }
